[Fix] Update crop_sub_images.py
Motivation:
There are several arguments in the script crop_sub_images.py should be int type, e.g., '--n-thread'. However, if we do not define them as int. They will be regarded as str if we pass those arguments to this script by ourselves, and therefore generate an error message.

Modification:
Define the int arguments with type=int explicitly.